fix(dev-overlay): move getIssueBucketState to standalone module to avoid cycle
