+++ packages/next/src/next-devtools/dev-overlay/menu/dev-overlay-menu.tsx
@@ -334,30 +334,6 @@ export function IssueCount({
   )
 }
 
-export type IssueBucketState = {
-  hasNormal: boolean
-  hasInstant: boolean
-  hasAny: boolean
-  insightsOnly: boolean
-  variant: 'issue' | 'insight'
-}
-
-export function getIssueBucketState(
-  normalErrorCount: number,
-  instantErrorCount: number
-): IssueBucketState {
-  const hasNormal = normalErrorCount > 0
-  const hasInstant = instantErrorCount > 0
-  const insightsOnly = !hasNormal && hasInstant
-  return {
-    hasNormal,
-    hasInstant,
-    hasAny: hasNormal || hasInstant,
-    insightsOnly,
-    variant: insightsOnly ? 'insight' : 'issue',
-  }
-}
-
 export function ChevronRight() {
   return (
     <svg
